1. Datenbank neu erstellen (alte migrations löschen, add-migration, update-database)
2. Region hinzufügen (FillRegionTable.sql) und Areas (FillAreaSizeTable.sql)
3. admin Benutzer erstellen. Usermail muss identisch sein mit einer in "Region", sonst kommt Error Meldung zurück das EMail in Region nicht vorhanden.
![image (3).png](https://dev.azure.com/silvanostecher/526ac74b-e0c7-4635-850a-f3668339b6e3/_apis/git/repositories/a38c64ec-3f32-47d4-9639-f0a34c045194/pullRequests/29/attachments/image%20%283%29.png)
4. Login und testen (Benutzer hinzufügen, anpassen, löschen)
5. Reponsive Design testen
6. Backend Tests sind angepasst
![image (4).png](https://dev.azure.com/silvanostecher/526ac74b-e0c7-4635-850a-f3668339b6e3/_apis/git/repositories/a38c64ec-3f32-47d4-9639-f0a34c045194/pullRequests/29/attachments/image%20%284%29.png)
![image (5).png](https://dev.azure.com/silvanostecher/526ac74b-e0c7-4635-850a-f3668339b6e3/_apis/git/repositories/a38c64ec-3f32-47d4-9639-f0a34c045194/pullRequests/29/attachments/image%20%285%29.png)

Related work items: #95

+++ RehkitzWebApp/clientapp/src/components/widgets/Protocol/Protocol.tsx
@@ -8,6 +8,7 @@ import ProtocolBodyLargeScreen from './ProtocolBodyLargeScreen'
 import { AppContext, ProtocolsContext } from '../../../store/context'
 import { ROUTE_ADAPT_PROTOCOL_PAGE } from '../../../App'
 import { useNavigate } from 'react-router-dom'
+import { toast } from 'react-toastify'
 
 interface Props {
     protocolId: String
@@ -15,7 +16,8 @@ interface Props {
 
 export default function Protocol({ protocolId }: Props) {
 
-    const isNotMobile = useMediaQuery({ query: '(min-width: 426px)' })
+    const isLargeScreen = useMediaQuery({ query: '(min-width: 1200px)' })
+
     //const { protocolId } = useParams() --> not working now as Router not set, there used Props
     const { token } = useContext(AppContext)
     const { protocolsListLocal, dispatch } = useContext(ProtocolsContext)
@@ -44,15 +46,22 @@ export default function Protocol({ protocolId }: Props) {
     }, [protocolsListLocal, protocolId]);
       
     const deleteProtocol = async (protocolId: string) => {
-        const response = await fetch(`/api/protocols/${Number(protocolId)}`, {
-            method: 'DELETE',
-            headers: { 
+        const answer = window.confirm("Wirklich löschen?")
+        if (answer) {
+            const response = await fetch(`/api/protocols/${Number(protocolId)}`, {
+                method: 'DELETE',
+                headers: { 
                 'content-type': 'application/json',
                 'Authorization': `Bearer ${token}`, // notice the Bearer before your token
-            },
-        })
-        if (response.ok) {
-            dispatch({ type: 'delete-protocols', protocolsListLocal, protocolId })
+                },
+            })
+            if (response.ok) {
+                dispatch({ type: 'delete-protocols', protocolsListLocal, protocolId })
+                toast.success("Protokoll erfolgreich gelöscht!", {
+                position: toast.POSITION.TOP_CENTER,
+                containerId: 'LoginToaster'
+                })
+            }
         }
     }
 
@@ -63,9 +72,9 @@ export default function Protocol({ protocolId }: Props) {
     return (
         <ProtocolLayout>
             <ProtocolTitle>Protokoll {protocolEntry.protocolCode}</ProtocolTitle>
-            {isNotMobile ? <ProtocolBodyLargeScreen protocolEntry={protocolEntry} /> : <ProtocolBodySmallScreen protocolEntry={protocolEntry} />}
+            {isLargeScreen ? <ProtocolBodyLargeScreen protocolEntry={protocolEntry} /> : <ProtocolBodySmallScreen protocolEntry={protocolEntry} />}
             <RowContainer>
-                <DeleteProtocolButton onClick={() => deleteProtocol(protocolEntry.protocolId)}>Loeschen</DeleteProtocolButton>
+                <DeleteProtocolButton onClick={() => deleteProtocol(protocolEntry.protocolId)}>Löschen</DeleteProtocolButton>
                 <EditProtocolButton onClick={() => editProtocol()}>Bearbeiten</EditProtocolButton>
             </RowContainer>
         </ProtocolLayout>
@@ -74,10 +83,11 @@ export default function Protocol({ protocolId }: Props) {
 
 const ProtocolLayout = styled.div`
 	margin: 10px;
+    padding: 20px;
 	display: flex;
 	flex-direction: column;
 	align-items: center;
-	background: #7d6b52;
+    background: #9A8873;
 	color: beige;
 	border-radius: 10px;
 `
@@ -86,6 +96,7 @@ const ProtocolTitle = styled.div`
 	font-weight: 500;
 	font-size: 25px;
 	margin: 10px;
+    color: #fffecb;
 `
 
 const RowContainer = styled.div`
